refactor(api): extract shared query parameter parsers
